source/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
django==3.2.8
djangorestframework==3.12.4
gunicorn==20.1.0
haversine==2.5.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
typing-extensions==3.10.0.2; python_version < '3.8'
